+++ src/store/reducer.js
@@ -1,5 +1,5 @@
 import { combineReducers } from 'redux';
-import authReducer from './auth';
+import authReducer from 'auth/store';
 import entitiesReducer from './entities';
 import messagesReducer from './messageSlice';
 import uiReducer from './uiSlice';